+++ R/linePlot_module_server.R
@@ -209,17 +209,10 @@ linePlotServer <- function(id, data, hide.inputs = NULL, hide.tabs = NULL, defau
 
 
         observeEvent(input$facet.by, {
-            if (.nz_value(input$facet.by)) {
-                show_input(session, c(
-                    "facet.title.font.size", "facet.title.font.color", "facet.title.font.family",
-                    "facet.nrow", "facet.ncol"
-                ))
-            } else {
-                hide_input(session, c(
-                    "facet.title.font.size", "facet.title.font.color", "facet.title.font.family",
-                    "facet.nrow", "facet.ncol"
-                ))
-            }
+            .toggle_facet_title_inputs(
+                session, .nz_value(input$facet.by),
+                extra = c("facet.nrow", "facet.ncol"), hidden = hide.inputs
+            )
         })
 
         # Reactive expression to generate the plot (used by both output and download)
@@ -368,6 +361,13 @@ linePlotServer <- function(id, data, hide.inputs = NULL, hide.tabs = NULL, defau
                 axis.tickwidth = isolate_fn(input$axis.tickwidth),
                 show.grid.x = isolate_fn(input$show.grid.x),
                 show.grid.y = isolate_fn(input$show.grid.y),
+                grid.color = isolate_fn(input$grid.color),
+                axis.title.font.size = isolate_fn(input$axis.title.font.size),
+                axis.title.font.color = isolate_fn(input$axis.title.font.color),
+                axis.title.font.family = isolate_fn(input$axis.title.font.family),
+                facet.title.font.size = isolate_fn(input$facet.title.font.size),
+                facet.title.font.color = isolate_fn(input$facet.title.font.color),
+                facet.title.font.family = isolate_fn(input$facet.title.font.family),
                 title.font.size = isolate_fn(input$title.font.size),
                 title.font.family = isolate_fn(input$title.font.family),
                 title.font.color = isolate_fn(input$title.font.color),
@@ -382,10 +382,6 @@ linePlotServer <- function(id, data, hide.inputs = NULL, hide.tabs = NULL, defau
                 error.width = isolate_fn(input$error.bar.width),
                 error.bar = isolate_fn(input$error.bar)
             )
-            # Apply axis title font to shared facet annotation titles
-            if (!is.null(facet.by) && nzchar(facet.by)) {
-                fig <- apply_axis_title_to_annotations(fig, input, isolate_fn)
-            }
             # Add reference lines
             fig <- add_reference_lines(fig,
                 hline.intercepts = isolate_fn(input$hline.intercepts),